fest: Rename the service's file for messages
