+++ server/api/browser/works/_workId@string/revisions/_revisionId@string/controller.ts
@@ -1,4 +1,4 @@
-import { createMessage, getMessages } from '$/service/stream'
+import { createMessage, getMessages } from '$/service/streamBar'
 import type { RevisionId } from '$/types'
 import { defineController } from './$relay'
