+++ server/service/streamBar.ts
